ingestion/requirements.txt
google-auth>=2.22.0
google-cloud-bigquery>=3.13.0
pandas>=2.0.3
pyarrow>=12.0.0
db-dtypes>=1.1.1
gspread>=5.10.0
google-cloud-storage>=2.10.0